+++ server/src/site/markdown/change-log.md
@@ -1,3 +1,7 @@
+1.4.1
+-------------
+* Fix webapp reports
+
 1.4.0
 -------------
 * Add Product [see](https://github.com/Axway/Grapes/wiki/Main-concepts#product)
@@ -34,5 +38,3 @@
 1.0.0
 -------------
 * Creation of Grapes project
-
-